Rewritten decoded text to be a static variable of the Decoder class
The program was previously appending decoded text to every Decoder class implementation instance's own decodedText variable. This was essentially a waste of time as each instance's decodedText was always set to ""; in other words, they did not keep track of the correct decoded text. By setting decodedText to be a static variable of the Decoder class, we can simly append new decoded text to it and keep track of the correct decoded text.

This commit also just creates instances of Decoder class implementations in the parseClasses array, since decodedText is a static variable now. It should also help performance in theory since ClassLooper no longer creates a new instance a Decoder class implementation when we want to decode an element of the raw array.